requirements.txt
Django==3.2.13
django-extensions==3.0.3
django-filter==2.3.0
django-jet2==1.0.12
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
django-better-admin-arrayfield==1.4.2
django-phonenumber-field==6.1.0
phonenumbers==8.12.48
psycopg2-binary==2.9.3
Pillow==9.1.1
social-auth-app-django==5.0.0
rest_condition==1.0.3